Add Myself (Carreau) as a maintainer
See https://pypi.org/project/pickleshare/, I can publish on PyPI. 

I'm mostly thinking of publishing a new version with a universal wheel, but maybe a bit of cleanup to avoid setup.py and only have a project.toml, so likely a new version number. 

Not right now, but I expect the new version (when I get to it), will be picked up by conda forge.